Direction change on implementation design
Rather than starting with a basic JDBC driver and growing from there, we will try to have a management layer on top of mysql's jdbc java connector.

This allows us to leave the mysql bits to the mysql driver and instead puts the complexities in deciding which driver to use, and when we can switch our underline connection.  These complexities would have to some extent existed in the other implementation eventually anyways.

There is still lots of 'TODO' notes, but this at least sets the framework for this design.