Fix / improvements by delaying delegate connection choice
Previously generating a Statement / PreparedStatement / CallableStatement was the point at which choosing the delegate connection was choosen.
This means that setting things like `readOnly` after preparing the statement were not able to be taken advantage of.  In addition cluster changes can not be applied while the rest of the statement is prepared / configured.
We now delay the connection choice longer, even queuing up actions / configuration until the point when we actually do generate these statements.
Right now that point is when any state is requested.  We don't try to interpet the state from previous settings, and instead just wait until some type of result needs to be returned.  This is simple and works for most cases.

This adds JDBI unit tests that previously would not route to the secondary server, but now which work as you might expect.