More javadocs and DelegatingConnection thread safety fixes
The previous change is not good enough.  We can not produce the delegator within the constructor.  Otherwise we wont be able to recognize possible read only switches.
However my previous attempt to proactively create the connection is not good enough because the creation still was not completely thread safe.  This new change fixes that, but in a brute synchronized way.  We probably will want to revist this and do something a bit better (with some benchmarking).