0.4 release with threadly 5.8 upgrade
As part of this upgrade we are switching to the new `CentralThreadlyPool` rather than having our own pool for monitoring the cluster.  This will hopefully reduce thread churn for applications using both.
The only down side to this is that we lost our priority semantics.  Before checking a potentially unhealthy server was done as a low priority task (as well as adding in read-slaves on startup).  Adding in read slaves is not so important, but the unhealthy server is a more interesting case.
I went back and improved the code where we are checking potentially unhealthy servers, solving the minor race condition that was there, so if multiple tasks get submitted the impact is lower.  This will alleviate the advantage we had by using low priority tasks there.