Support for postgresql driver
This resolves #15 in providing both a mysql and psql driver.

The auroraArc artifact of the past has been replaced by `auroraArc-mysql` and `auroraArc-psql` to let you specify what delegate driver implementation you want to use.

Gradle is sort of weird with specifying the `-b` flag when using sub-projects.  I could not find a way to solve it, so for now this has a `.sh` file that will replace the file to do the upload.