The fixes psql support
This resolves #18 by correctly handling the postgresql specific master / read only logic aurora provides.
If strange cname like dns is attempted to be used, this current solution will not work.  It does require the host URL's to be exactly as AWS provides them.  In the future when detecting cluster membership is done this may not be an issue.