DelegatingAuroraConnection: Allow more control on delegate choice
This allows you to specify more precise what type of logic should be used in picking a delegate connection.
Previously the only control was through setting the readOnly status.  This behavior is still defaulted as the `Smart` option.
But in addition behaviors to use only the master, or only a secondary, or even distribute to specific secondary servers have been added.
This can allow for more performant control over which connections to use, but can also allow options to make better use of the cluster.

Please read the javadocs at the top of `DelegatingAuroraConnection` for more details on the different behaviors this now provides.