Added AuroraClusterMonitor.setServerCheckDelayMillis
This solves a TODO note to allow the frequency at which the server master / slave status is checked.  Though the default of 2 a second should be good in most cases, I can see where a small client count system may want to reduce this further, or where a larger service may want to increase the delay.

Between this commit and the last commit (which introduced setClientInfo configuration), this resolves #8