DelegatingAuroraConnection: Delay connect failures till connection is used
This changes DelegatingAuroraConnection so that as long as at least one connection works, failures to connect wont propagate out of the constructor.
Instead the invalid connection will appear valid until it is attempted to be used, at which point an exception will be thrown with the original failure as the cause.

This is to improve the ability to use the driver and establish new connections while the aurora cluster is partially unhealthy.
Without this Hikari (or other pools) may end up fully removing the connections from the pool as they appear to be invalid / unhealthy, and be also unable to replace them with new connections, even though the aurora cluster is partially usable.

This does NOT delay errors when starting up the AuroraClusterMonitor, so starting up a service and needing to monitor a cluster that is partially unhealthy still wont work.

This is better than lazily establishing the connections because we don't want the connection establish delay to occur at the time of the operation / request (rather have the overhead while the pool is filling / replenishing).