AuroraClusterMonitor: Allow weighted replica's
This adds the static function AuroraClusterMonitor.setReplicaWeight(String, int, int) to be able to set weights for a given server.
As described in the javadocs, the weight effectively represents how many times that server should be choosen as a random replica among the cluster.
A weight of zero will prevent the server from being used unless it's the only available healthy replica.  While larger numbers will deliever additional traffic to those instances.

In addition dependency versions were updated and JMH was added as a test dependency.  A lot of benchmarking was done to make sure that this does not impact the performance of the driver.

In order to avoid performance impacts the weights are managed in additoinal collection.  Making the cluster state checker do a little additional logic to manage this collection.  Allowing the server selection logic to remain almost unchanged, and with no statistically significant performance impact.