Changes to allow "Any" server to consider weights
This change moves the logic for choosing "Any" server from DelegatingAuroraConnection into the AuroraClusterMonitor.
This allowed us to consider weights for the master server against replica's, while still failing over to a zero weight server if no weighted server is available.

In order to do this, we extend the previous weight implementation to now also manage a list of all servers.